requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==3.0.4
colorama==0.4.6
Django==5.1.5
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
flake8==7.1.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.4
pytest-django==4.9.0
redis==5.2.1
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.1
